+++ src/ConsensusCore/src/C++/Matrix/DenseMatrix.hpp
@@ -93,17 +93,17 @@ namespace ConsensusCore {
         // is a set or get.  To my knowledge, there is no way to do this in
         // C++.
         //
-        const float& operator()(int i, int j) const;
+        const double& operator()(int i, int j) const;
         bool IsAllocated(int i, int j) const;
-        float Get(int i, int j) const;
-        void Set(int i, int j, float v);
+        double Get(int i, int j) const;
+        void Set(int i, int j, double v);
         void ClearColumn(int j);
 
 
     public:
         // Method SWIG clients can use to get a native matrix (e.g. Numpy)
         // mat must be filled as a ROW major matrix
-        void ToHostMatrix(float** mat, int* rows, int* cols) const;
+        void ToHostMatrix(double** mat, int* rows, int* cols) const;
 
     private:
         std::vector<Interval> usedRanges_;
